fix(components, modal): use forwardRef
