+++ packages/components/src/modal/index.ts
@@ -1,2 +1,2 @@
-export { Modal, DialogTrigger, type ModalProps } from './Dialog'
+export { Modal, DialogTrigger, type ModalProps } from './Modal'
 export * from './modal-overlay'
